+++ plugin-gradle/src/test/java/com/diffplug/gradle/spotless/SpotlessPluginRedirectTest.java
@@ -15,13 +15,10 @@
  */
 package com.diffplug.gradle.spotless;
 
-import static org.junit.jupiter.api.condition.JRE.JAVA_11;
-
 import java.io.IOException;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.condition.EnabledForJreRange;
 
 import com.diffplug.common.base.StringPrinter;
 
@@ -38,36 +35,5 @@ class SpotlessPluginRedirectTest extends GradleIntegrationHarness {
 						"                     to 'com.diffplug.spotless'",
 						"     To migrate:",
 						"     - Test your build with: id 'com.diffplug.gradle.spotless' version '4.5.1'"));
-	}
-
-	@Test
-	@EnabledForJreRange(max = JAVA_11)
-	void redirectPluginOldGradle() throws IOException {
-		setFile("build.gradle").toLines(
-				"plugins {",
-				"    id 'com.diffplug.gradle.spotless'",
-				"}");
-		Assertions.assertThat(gradleRunner().withGradleVersion(GradleVersionSupport.JRE_11.version)
-				.buildAndFail().getOutput().replace("\r", ""))
-				.contains(StringPrinter.buildStringFromLines(
-						"> Failed to apply plugin [id 'com.diffplug.gradle.spotless']",
-						"   > We have moved from 'com.diffplug.gradle.spotless'",
-						"                     to 'com.diffplug.spotless'",
-						"     To migrate:",
-						"     - Upgrade Gradle to 7.1 or newer (you're on 5.0)",
-						"     - Test your build with: id 'com.diffplug.gradle.spotless' version '4.5.1'"));
-	}
-
-	@Test
-	@EnabledForJreRange(max = JAVA_11)
-	void realPluginOldGradle() throws IOException {
-		setFile("build.gradle").toLines(
-				"plugins {",
-				"    id 'com.diffplug.spotless'",
-				"}");
-		Assertions.assertThat(gradleRunner().withGradleVersion(GradleVersionSupport.JRE_11.version)
-				.buildAndFail().getOutput().replace("\r", ""))
-				.contains(StringPrinter.buildStringFromLines(
-						"Spotless requires Gradle 7.1 or newer, this was 5.0"));
 	}
 }